requirements.txt
arrow==0.15.7
asgiref==3.2.10
beautifulsoup4==4.9.1
binaryornot==0.4.4
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
dj-database-url==0.5.0
Django==3.0.7
django-async-include==0.6.5
django-cors-headers==3.5.0
django-datetimepicker==3.14
django-heroku==0.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.2
jinja2-time==0.2.0
jsonpickle==1.4.1
lxml==4.5.2
MarkupSafe==1.1.1
Pillow==7.2.0
poyo==0.5.0
psycopg2==2.8.5
pycryptodome==3.9.8
python-dateutil==2.8.1
python-dotenv==0.14.0
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.1.0
zipp==3.1.0
